pyproject.toml
[tool.poetry]
name = "arduino-sensor-streamlit"
version = "0.1.0"
description = ""
authors = ["Aditya Wardianto <adityawardianto.19073@mhs.its.ac.id>"]

[tool.poetry.dependencies]
python = "3.9.7"
streamlit = "^1.2.0"
pyserial = "^3.5"
plotly = "^5.4.0"

[tool.poetry.dev-dependencies]
black = "^21.12b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
